Hide dep guts under internal/
* Move dep/ to internal/
* Move testdata/ to internal/testdata
* Move log funcs into own internal package so test, internal and cmd/dep can reference.
  Otherwise circular references were present.
  Using the name util in honor of @kris-nova and to ensure proper bike-shedding.
* Move test/ into internal/test/ so that internal tests can use it
