+++ cmd/dep/integration_test.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/golang/dep"
-	"github.com/golang/dep/test"
+	"github.com/golang/dep/internal"
+	"github.com/golang/dep/internal/test"
 )
 
 func TestIntegration(t *testing.T) {
@@ -68,8 +68,8 @@ func TestIntegration(t *testing.T) {
 				testCase.CompareError(err, testProj.GetStderr())
 
 				// Check final manifest and lock
-				testCase.CompareFile(dep.ManifestName, testProj.ProjPath(dep.ManifestName))
-				testCase.CompareFile(dep.LockName, testProj.ProjPath(dep.LockName))
+				testCase.CompareFile(internal.ManifestName, testProj.ProjPath(internal.ManifestName))
+				testCase.CompareFile(internal.LockName, testProj.ProjPath(internal.LockName))
 
 				// Check vendor paths
 				testProj.CompareImportPaths()
