+++ cmd/dep/remove.go
@@ -10,10 +10,10 @@ import (
 	"os"
 	"strings"
 
-	"github.com/golang/dep"
 	"github.com/golang/dep/gps"
 	"github.com/golang/dep/gps/pkgtree"
 	"github.com/golang/dep/internal"
+	"github.com/golang/dep/internal/util"
 	"github.com/pkg/errors"
 )
 
@@ -44,7 +44,7 @@ type removeCommand struct {
 	keepSource bool
 }
 
-func (cmd *removeCommand) Run(ctx *dep.Ctx, args []string) error {
+func (cmd *removeCommand) Run(ctx *internal.Ctx, args []string) error {
 	p, err := ctx.LoadProject("")
 	if err != nil {
 		return err
@@ -92,7 +92,7 @@ func (cmd *removeCommand) Run(ctx *dep.Ctx, args []string) error {
 				// not being able to detect the root for an import path that's
 				// actually in the import list is a deeper problem. However,
 				// it's not our direct concern here, so we just warn.
-				internal.Logf("could not infer root for %q", pr)
+				util.Logf("could not infer root for %q", pr)
 				continue
 			}
 			otherroots[pr] = true
@@ -107,7 +107,7 @@ func (cmd *removeCommand) Run(ctx *dep.Ctx, args []string) error {
 		}
 
 		if len(rm) == 0 {
-			internal.Logf("nothing to do")
+			util.Logf("nothing to do")
 			return nil
 		}
 	} else {
@@ -180,9 +180,9 @@ func (cmd *removeCommand) Run(ctx *dep.Ctx, args []string) error {
 		return err
 	}
 
-	newLock := dep.LockFromInterface(soln)
+	newLock := internal.LockFromInterface(soln)
 
-	sw, err := dep.NewSafeWriter(nil, p.Lock, newLock, dep.VendorOnChanged)
+	sw, err := internal.NewSafeWriter(nil, p.Lock, newLock, internal.VendorOnChanged)
 	if err != nil {
 		return err
 	}
