+++ internal/analyzer_test.go
@@ -2,13 +2,13 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package dep
+package internal
 
 import (
 	"path/filepath"
 	"testing"
 
-	"github.com/golang/dep/test"
+	"github.com/golang/dep/internal/test"
 )
 
 func TestAnalyzerDeriveManifestAndLock(t *testing.T) {
