+++ internal/lock_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package dep
+package internal
 
 import (
 	"encoding/hex"
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/golang/dep/gps"
-	"github.com/golang/dep/test"
+	"github.com/golang/dep/internal/test"
 )
 
 func TestReadLock(t *testing.T) {
