+++ internal/testdata/lock/golden0.toml
