+++ internal/testdata/manifest/golden.toml
