+++ internal/testdata/txn_writer/expected_diff_output.txt
